conversion fix in Wall::Render()

+++ Source/game.cpp
@@ -734,7 +734,7 @@ void Wall::Render(Texture2D texture) const
 		WHITE);
 
 
-	DrawText(TextFormat("%i", health), position.x-21, position.y+10, 40, RED);
+	DrawText(TextFormat("%i", health), (int)position.x-21, (int)position.y+10, 40, RED);
 	
 }
